# go/chart-inspector/docs/swagger.json
{"swagger":"2.0","info":{"description":"This is the API for the Chart Inspector service. It provides endpoints for inspecting Helm charts.","title":"Chart Inspector API","contact":{},"version":"1.0"},"basePath":"/","paths":{"/resources":{"get":{"description":"Get Helm chart resources","produces":["application/json"],"summary":"Get Helm chart resources","operationId":"get-chart-resources","parameters":[{"type":"string","description":"Composition name","name":"compositionName","in":"query","required":true},{"type":"string","description":"Composition namespace","name":"compositionNamespace","in":"query","required":true},{"type":"string","description":"Composition definition name","name":"compositionDefinitionName","in":"query","required":true},{"type":"string","description":"Composition definition namespace","name":"compositionDefinitionNamespace","in":"query","required":true},{"type":"string","default":"core.krateo.io","description":"Composition definition group","name":"compositionDefinitionGroup","in":"query"},{"type":"string","default":"v1alpha1","description":"Composition definition version","name":"compositionDefinitionVersion","in":"query"},{"type":"string","default":"compositiondefinitions","description":"Composition definition resource name","name":"compositionDefinitionResource","in":"query"},{"type":"string","default":"composition.krateo.io","description":"Composition group","name":"compositionGroup","in":"query"},{"type":"string","description":"Composition version","name":"compositionVersion","in":"query","required":true},{"type":"string","description":"Composition resource name","name":"compositionResource","in":"query","required":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/resources.Resource"}}}}}}},"definitions":{"resources.Resource":{"type":"object","properties":{"group":{"type":"string"},"name":{"type":"string"},"namespace":{"type":"string"},"resource":{"type":"string"},"version":{"type":"string"}}}}}